+++ R/hetcalc.R
@@ -1,7 +1,7 @@
 hetcalc <- function(TE, seTE,
                     method.tau, method.tau.ci,
                     TE.tau, level.hetstats, subgroup, control,
-                    id = NULL) {
+                    cluster = NULL) {
   
   Ccalc <- function(x) {
     res <- (sum(x, na.rm = TRUE) -
@@ -17,16 +17,16 @@ hetcalc <- function(TE, seTE,
   sel.noInf <- !is.infinite(TE) & !is.infinite(seTE)
   TE <- TE[sel.noInf]
   seTE <- seTE[sel.noInf]
-  if (!is.null(id))
-    id <- id[sel.noInf]
+  if (!is.null(cluster))
+    cluster <- cluster[sel.noInf]
   if (by)
     subgroup <- subgroup[sel.noInf]
   ##
   sel.noNA <- !(is.na(TE) | is.na(seTE))
   TE <- TE[sel.noNA]
   seTE <- seTE[sel.noNA]
-  if (!is.null(id))
-    id <- id[sel.noNA]
+  if (!is.null(cluster))
+    cluster <- cluster[sel.noNA]
   if (by)
     subgroup <- subgroup[sel.noNA]
   ##
@@ -79,7 +79,7 @@ hetcalc <- function(TE, seTE,
       sign.lower.tau <- sign.upper.tau <- method.tau.ci <- ""
     }
     else {
-      if (is.null(id)) {
+      if (is.null(cluster)) {
         mf0 <- runNN(rma.uni,
                      list(yi = TE, sei = seTE, method = method.tau,
                           control = control))
@@ -95,9 +95,9 @@ hetcalc <- function(TE, seTE,
         mf0 <-
           runNN(rma.mv,
                 list(yi = TE, V = seTE^2, method = method.tau,
-                     random = as.call(~ 1 | id / idx),
+                     random = as.call(~ 1 | cluster / idx),
                      control = control,
-                     data = data.frame(id, idx)),
+                     data = data.frame(cluster, idx)),
                 warn = FALSE)
         ##
         tau2 <- mf0$sigma2
@@ -113,7 +113,7 @@ hetcalc <- function(TE, seTE,
       ##
       if (df.Q < 2)
         method.tau.ci <- ""
-      else if (!is.null(id) & method.tau.ci != "")
+      else if (!is.null(cluster) & method.tau.ci != "")
         method.tau.ci <- "PL"
       ##
       ## Confidence interal for overall result
@@ -147,7 +147,7 @@ hetcalc <- function(TE, seTE,
     if (is.numeric(subgroup))
       subgroup <- as.factor(subgroup)
     ##
-    if (is.null(id)) {
+    if (is.null(cluster)) {
       if (length(unique(subgroup)) == 1)
         mf1 <-
           runNN(rma.uni,
@@ -186,18 +186,18 @@ hetcalc <- function(TE, seTE,
         mf1 <-
           runNN(rma.mv,
                 list(yi = TE, V = seTE^2, method = method.tau,
-                     random = as.call(~ 1 | id / idx),
+                     random = as.call(~ 1 | cluster / idx),
                      control = control,
-                     data = data.frame(id, idx)),
+                     data = data.frame(cluster, idx)),
                 warn = FALSE)
       else {
         mf1 <-
           try(
             runNN(rma.mv,
                   list(yi = TE, V = seTE^2, method = method.tau,
-                       random = as.call(~ 1 | id / idx),
+                       random = as.call(~ 1 | cluster / idx),
                        mods = as.call(~ subgroup), control = control,
-                       data = data.frame(TE, seTE, subgroup, id, idx)),
+                       data = data.frame(TE, seTE, subgroup, cluster, idx)),
                   warn = FALSE),
             silent = TRUE)
         ##
@@ -210,9 +210,9 @@ hetcalc <- function(TE, seTE,
               runNN(rma.mv,
                     list(yi = TE, V = seTE^2,
                          method = "FE",
-                         random = as.call(~ 1 | id / idx),
+                         random = as.call(~ 1 | cluster / idx),
                          mods = as.call(~ subgroup), control = control,
-                         data = data.frame(TE, seTE, subgroup, id, idx)),
+                         data = data.frame(TE, seTE, subgroup, cluster, idx)),
                     warn = FALSE)
           }
           else
@@ -231,7 +231,7 @@ hetcalc <- function(TE, seTE,
     ##
     if (df.Q < 2 || useFE)
       method.tau.ci <- ""
-    else if (!is.null(id) & method.tau.ci != "")
+    else if (!is.null(cluster) & method.tau.ci != "")
       method.tau.ci <- "PL"
     ##
     ## Confidence interal for residual heterogeneity